hypergrid.js cleanup.
