+++ packages/perspective-viewer-hypergrid/src/js/hypergrid.js
@@ -7,6 +7,8 @@
  *
  */
 
+import {GridUIFixPlugin} from "./fixes.js";
+
 const Hypergrid = require("fin-hypergrid");
 const Behaviors = require("fin-hypergrid/src/behaviors");
 const Base = require("fin-hypergrid/src/Base.js");
@@ -15,8 +17,6 @@ const treeLineRendererPaint = require("./hypergrid-tree-cell-renderer.js").treeL
 const GroupedHeader = require("./grouped-header.js");
 
 const _ = require("underscore");
-
-const rectangular = require('rectangular')
 
 import {registerElement, detectChrome} from "@jpmorganchase/perspective-common";
 
@@ -117,7 +117,7 @@ function generateGridProperties(overrides) {
 }
 
 function setPSP(payload) {
-    if (payload.data.length === 0) {
+    if (payload.rows.length === 0) {
         this.grid.setData({data: []});
         return;
     };
@@ -149,11 +149,9 @@ function setPSP(payload) {
     var old_schema = this.grid.behavior.subgrids.lookup.data.schema;
     this.schema_loaded = this.schema_loaded && _.isEqual(processed_schema, old_schema);
     this.schema = processed_schema;
-    var mergedData = mergePathsAndRows(payload.rowPaths, payload.data, payload.rowLeaf);
-    this.grid._rowPathsIndex = mergedData.rowPathsIndex;
     if (this.schema_loaded) {
         this.grid.setData({
-            data: mergedData.data,
+            data: payload.rows,
         });
     } else {
 
@@ -172,7 +170,7 @@ function setPSP(payload) {
         }
         console.log('Setting up initial schema and data load into HyperGrid');
         this.grid.behavior.setData({
-            data: mergedData.data,
+            data: payload.rows,
             schema: this.schema
         });
         this.schema_loaded = true;
@@ -227,315 +225,6 @@ function setPSP(payload) {
 
 }
 
-function GridUIFixPlugin(grid) {
-
-    grid.canvas.resize = function() {
-        var box = this.size = this.div.getBoundingClientRect();
-
-        this.width = Math.floor(this.div.clientWidth);
-        this.height = Math.floor(this.div.clientHeight);
-
-        //fix ala sir spinka, see
-        //http://www.html5rocks.com/en/tutorials/canvas/hidpi/
-        //just add 'hdpi' as an attribute to the fin-canvas tag
-        var ratio = 1;
-        var isHIDPI = window.devicePixelRatio && this.component.properties.useHiDPI;
-        if (isHIDPI) {
-            var devicePixelRatio = window.devicePixelRatio || 1;
-            var backingStoreRatio = this.gc.webkitBackingStorePixelRatio ||
-                this.gc.mozBackingStorePixelRatio ||
-                this.gc.msBackingStorePixelRatio ||
-                this.gc.oBackingStorePixelRatio ||
-                this.gc.backingStorePixelRatio || 1;
-
-            ratio = devicePixelRatio / backingStoreRatio;
-            //this.canvasCTX.scale(ratio, ratio);
-        }
-
-        this.buffer.width = this.canvas.width = this.width * ratio;
-        this.buffer.height = this.canvas.height = this.height * ratio;
-
-        this.canvas.style.width = this.buffer.style.width = this.width + 'px';
-        this.canvas.style.height = this.buffer.style.height = this.height + 'px';
-
-        this.bc.scale(ratio, ratio);
-        if (isHIDPI && !this.component.properties.useBitBlit) {
-            this.gc.scale(ratio, ratio);
-        }
-
-        this.bounds = new rectangular.Rectangle(0, 0, this.width, this.height);
-        this.component.setBounds(this.bounds);
-        this.resizeNotification();
-        this.paintNow();
-    }
-
-    grid.canvas._tickPaint = grid.canvas.tickPaint;
-    grid.canvas.tickPaint = async function (t) {
-        let range = this.component.grid.getVisibleRows();
-        let s = range[1];
-        let e = range[range.length - 1];
-        if (range.length > 1 && (this.dirty || this.__cached_start !== s || this.__cached_end !== e)) {
-            if (this._updating_cache) {
-                this._updating_cache.cancel();
-            }
-            this._updating_cache = this.component.grid._cache_update(s, e);
-            await this._updating_cache;
-            this._updateing_cache = undefined;
-            this.__cached_start = s;
-            this.__cached_end = e;
-        }
-        this.component.grid.canvas._tickPaint(t);
-    }
-
-    grid._getGridCellFromMousePoint = grid.getGridCellFromMousePoint;
-    grid.getGridCellFromMousePoint = function(mouse) {
-        if (this.getRowCount() === 0) {
-            return {'fake': 1};
-        } else {
-            try {
-                return this._getGridCellFromMousePoint(mouse);
-            } catch (e) {
-                return {'fake': 1};
-
-            }
-        }
-    };
-
-
-    grid.renderer._paintGridlines = grid.renderer.paintGridlines;
-    grid.renderer.paintGridlines = function(gc) {
-        var visibleColumns = this.visibleColumns, C = visibleColumns.length,
-            visibleRows = this.visibleRows, R = visibleRows.length;
-
-        if (C && R) {
-            var gridProps = this.properties,
-                lineWidth = gridProps.lineWidth,
-                lineColor = gridProps.lineColor,
-                viewHeight = visibleRows[R - 1].bottom,
-                viewWidth = visibleColumns[C - 1].right;
-
-            gc.cache.fillStyle = lineColor;
-
-            if (gridProps.fixedColumnCount > 0 && gridProps.fixedColumnCount < C) {
-                var lw = gridProps.gridLinesV ? lineWidth + 1: lineWidth;
-                var fixedX = visibleColumns[gridProps.fixedColumnCount].left - 1;
-                gc.fillRect(fixedX, 0, lw, viewHeight);
-            }
-
-            if (gridProps.fixedRowCount > 0 && gridProps.fixedRowCount < R) {
-                var lw = gridProps.gridLinesH ? lineWidth + 1: lineWidth;
-                var fixedY = visibleRows[gridProps.fixedRowCount].bottom;
-                gc.fillRect(0, fixedY, viewWidth, lw);
-            }
-
-            gc.fillRect(0, visibleRows[0].bottom, viewWidth, 1);
-
-            this._paintGridlines(gc);
-        }
-    };
-
-    grid.renderer.computeCellsBounds = function() {
-
-        var scrollTop = this.getScrollTop(),
-            scrollLeft = this.getScrollLeft(),
-
-            fixedColumnCount = this.grid.getFixedColumnCount(),
-            fixedRowCount = this.grid.getFixedRowCount(),
-
-            bounds = this.getBounds(),
-            grid = this.grid,
-            behavior = grid.behavior,
-            editorCellEvent = grid.cellEditor && grid.cellEditor.event,
-
-            vcEd, xEd,
-            vrEd, yEd,
-            sgEd, isSubgridEd,
-
-            insertionBoundsCursor = 0,
-            previousInsertionBoundsCursorValue = 0,
-
-            lineWidthX = grid.properties.gridLinesV ? grid.properties.lineWidth : 1,
-            lineWidthY = grid.properties.gridLinesH ? grid.properties.lineWidth : 0,
-
-            start = 0,
-            numOfInternalCols = 0,
-            x, X, // horizontal pixel loop index and limit
-            y, Y, // vertical pixel loop index and limit
-            c, C, // column loop index and limit
-            g, G, // subgrid loop index and limit
-            r, R, // row loop index and limitrows in current subgrid
-            subrows, // rows in subgrid g
-            base, // sum of rows for all subgrids so far
-            subgrids = behavior.subgrids,
-            subgrid,
-            rowIndex,
-            scrollableSubgrid,
-            footerHeight,
-            vx, vy,
-            vr, vc,
-            width, height,
-            firstVX, lastVX,
-            firstVY, lastVY,
-            topR,
-            xSpaced, widthSpaced, heightSpaced; // adjusted for cell spacing
-
-        if (editorCellEvent) {
-            xEd = editorCellEvent.gridCell.x;
-            yEd = editorCellEvent.dataCell.y;
-            sgEd = editorCellEvent.subgrid;
-        }
-        if (grid.properties.showRowNumbers) {
-            start = behavior.rowColumnIndex;
-            numOfInternalCols = Math.abs(start);
-        }
-
-        this.scrollHeight = 0;
-
-        this.visibleColumns.length = 0;
-        this.visibleRows.length = 0;
-
-        this.visibleColumnsByIndex = []; // array because number of columns will always be reasonable
-        this.visibleRowsByDataRowIndex = {}; // hash because keyed by (fixed and) scrolled row indexes
-
-        this.insertionBounds = [];
-
-        for (
-            x = 0, c = start, C = grid.getColumnCount(), X = bounds.width || grid.canvas.width;
-            c < C && x <= X;
-            c++
-        ) {
-            if (c === behavior.treeColumnIndex && !behavior.hasTreeColumn()) {
-                numOfInternalCols = (numOfInternalCols > 0) ? numOfInternalCols - 1 : 0;
-                this.visibleColumns[c] = undefined;
-                continue;
-            }
-
-            vx = c;
-            if (c >= fixedColumnCount) {
-                lastVX = vx += scrollLeft;
-                if (firstVX === undefined) {
-                    firstVX = lastVX;
-                }
-            }
-            if (vx >= C) {
-                break; // scrolled beyond last column
-            }
-
-            width = Math.ceil(behavior.getColumnWidth(vx));
-
-            xSpaced = x ? x + lineWidthX : x;
-            widthSpaced = x ? width - lineWidthX : width;
-            this.visibleColumns[c] = this.visibleColumnsByIndex[vx] = vc = {
-                index: c,
-                columnIndex: vx,
-                column: behavior.getActiveColumn(vx),
-                left: xSpaced,
-                width: widthSpaced,
-                right: xSpaced + widthSpaced
-            };
-            if (xEd === vx) {
-                vcEd = vc;
-            }
-
-            x += width;
-
-            insertionBoundsCursor += Math.round(width / 2) + previousInsertionBoundsCursorValue;
-            this.insertionBounds.push(insertionBoundsCursor);
-            previousInsertionBoundsCursorValue = Math.round(width / 2);
-        }
-
-        // get height of total number of rows in all subgrids following the data subgrid
-        footerHeight = grid.properties.defaultRowHeight *
-            subgrids.reduce(function(rows, subgrid) {
-                if (scrollableSubgrid) {
-                    rows += subgrid.getRowCount();
-                } else {
-                    scrollableSubgrid = subgrid.isData;
-                }
-                return rows;
-            }, 0);
-
-        for (
-            base = r = g = y = 0, G = subgrids.length, Y = bounds.height - footerHeight;
-            g < G;
-            g++, base += subrows
-        ) {
-            subgrid = subgrids[g];
-            subrows = subgrid.getRowCount();
-            scrollableSubgrid = subgrid.isData;
-            isSubgridEd = (sgEd === subgrid);
-            topR = r;
-
-            // For each row of each subgrid...
-            for (R = r + subrows; r < R && y < Y; r++) {
-                vy = r;
-                if (scrollableSubgrid && r >= fixedRowCount) {
-                    vy += scrollTop;
-                    lastVY = vy - base;
-                    if (firstVY === undefined) {
-                        firstVY = lastVY;
-                    }
-                    if (vy >= R) {
-                        break; // scrolled beyond last row
-                    }
-                }
-
-                rowIndex = vy - base;
-                height = behavior.getRowHeight(rowIndex, subgrid);
-
-                heightSpaced = height - lineWidthY;
-                this.visibleRows[r] = vr = {
-                    index: r,
-                    subgrid: subgrid,
-                    rowIndex: rowIndex,
-                    top: y,
-                    height: heightSpaced,
-                    bottom: y + heightSpaced
-                };
-
-                if (scrollableSubgrid) {
-                    this.visibleRowsByDataRowIndex[vy - base] = vr;
-                }
-
-                if (isSubgridEd && yEd === rowIndex) {
-                    vrEd = vr;
-                }
-
-                y += height;
-            }
-
-            if (scrollableSubgrid) {
-                subrows = r - topR;
-                Y += footerHeight;
-            }
-        }
-
-        if (editorCellEvent) {
-            editorCellEvent.visibleColumn = vcEd;
-            editorCellEvent.visibleRow = vrEd;
-            editorCellEvent.gridCell.y = vrEd && vrEd.index;
-            editorCellEvent._bounds = null;
-        }
-
-        this.viewHeight = Y;
-
-        this.dataWindow = this.grid.newRectangle(firstVX, firstVY, lastVX - firstVX, lastVY - firstVY);
-
-        // Resize CellEvent pool
-        var pool = this.cellEventPool,
-            previousLength = pool.length,
-            P = (this.visibleColumns.length + numOfInternalCols) * this.visibleRows.length;
-
-        if (P > previousLength) {
-            pool.length = P; // grow pool to accommodate more cells
-        }
-        for (var p = previousLength; p < P; p++) {
-            pool[p] = new behavior.CellEvent; // instantiate new members
-        }
-
-        this.resetAllGridRenderers();
-    };
-}
 
 function CheckboxTrackingPlugin(grid) {
 
@@ -565,7 +254,11 @@ function CheckboxTrackingPlugin(grid) {
 
     grid.getRowIdx = function (rowPath) {
         var path = Array.isArray(rowPath) ? JSON.stringify(rowPath) : rowPath;
-        return grid._rowPathsIndex[path];
+        for (let i = 0; i < grid.getRowCount(); i ++) {
+            if (JSON.stringify(grid.getRow(i).rowPath) === path) {
+                return i;
+            }
+        }
     };
 
 }
@@ -584,7 +277,7 @@ function PerspectiveDataModel(grid) {
     grid.mixIn.call(grid.behavior.dataModel, {
 
         // Override setData
-        setData: function (dataPayload, schema, cache_update) {
+        setData: function (dataPayload, schema) {
             this.viewData = dataPayload;
             this.source.setData(dataPayload, schema);
         },
@@ -634,14 +327,6 @@ function PerspectiveDataModel(grid) {
 
         // Returns the number of rows for this dataset
         getRowCount: function () {
-            // let range = this.grid.getVisibleRows();
-            // let s = range[1];
-            // let e = range[range.length - 1];
-            // if (range.length > 1 && (this.__cached_start !== s || this.__cached_end !== e)) {
-            //     this.__cached_start = s;
-            //     this.__cached_end = e;
-            //     this.grid._cache_update(s, e);
-            // }
             return this.dataSource.data.length;
         },
 
@@ -725,19 +410,6 @@ function PerspectiveDataModel(grid) {
     });
 }
 
-
-function mergePathsAndRows(rowPaths, rowData, rowLeaf) {
-    var mergedData = [];
-    var rowPathsIndex = {};
-    for (var i = 0; i < rowData.length; i++) {
-        var row_data = { rowPath: rowPaths[i], rowData: rowData[i] };
-        row_data['isLeaf'] = rowLeaf ? rowLeaf[i] : true;
-        mergedData.push(row_data);
-        rowPathsIndex[JSON.stringify(rowPaths[i])] = i;
-    }
-    return { data: mergedData, rowPathsIndex: rowPathsIndex };
-}
-
 function convertToType(typ, val) {
     return ['object', 'boolean'].indexOf(typeof (typ)) > -1 ? JSON.parse(val) : (typ.constructor)(val);
 }
@@ -761,6 +433,7 @@ function psp2hypergrid(data, schema) {
             columnTypes: []
         }
     }
+
     var is_tree = data[0].hasOwnProperty('__ROW_PATH__');
 
     var columnPaths = Object.keys(data[0])
@@ -769,42 +442,39 @@ function psp2hypergrid(data, schema) {
 
     let flat_columns = columnPaths.map(col => col.join(","));
 
-    let row_paths = [];
     let rows = [];
-    let row_leaves = [];
     for (let idx = 0; idx < data.length; idx++) {
         const row = data[idx] || {};
         let new_row = [];
+        let row_path = [];
+        let row_leaf = true;
         if (is_tree) {
             if (row.__ROW_PATH__ === undefined) {
                 row.__ROW_PATH__ = [];
             }
-            row_paths.push(["ROOT"].concat(row.__ROW_PATH__));
+            row_path = ["ROOT"].concat(row.__ROW_PATH__);
             let name = row['__ROW_PATH__'][row['__ROW_PATH__'].length - 1];
             if (name === undefined && idx === 0) name = "TOTAL"
             new_row = [name];
-            row_leaves.push(row.__ROW_PATH__.length >= (data[idx + 1] ? data[idx + 1].__ROW_PATH__.length : 0));
-        } else {
-            row_paths.push([]);
+            row_leaf = row.__ROW_PATH__.length >= (data[idx + 1] ? data[idx + 1].__ROW_PATH__.length : 0);
         }
         for (var col of flat_columns) {
             new_row.push(row[col]);
         }
-        rows.push(new_row);
+        rows.push({
+            rowPath: row_path,
+            rowData: new_row,
+            rowLeaf: row_leaf
+        });
     }
 
     var hg_data = {
-        rowPaths: row_paths,
-        data: rows,
+        rows: rows,
         isTree: is_tree,
         configuration: {},
         columnPaths: (is_tree ? [[" "]] : []).concat(columnPaths),
         columnTypes: (is_tree ? ["str"] : []).concat(columnPaths.map(col => conv[schema[col[col.length - 1]]]))
     };
-
-    if (is_tree) {
-        hg_data['rowLeaf'] = row_leaves;
-    }
 
     return hg_data
 }
@@ -917,16 +587,6 @@ async function fill_page(view, json, hidden, start_row, end_row) {
     return json;
 }
 
-async function load_incrementally(view, schema, hidden, json, total, page) {
-    json = await fill_page(view, json, hidden, page * PAGE_SIZE, (page + 1) * PAGE_SIZE);
-    this.grid.set_data(json, schema);
-    this.grid.grid.canvas.resize();
-    this.grid.grid.canvas.resize();
-    if ((page + 1) * PAGE_SIZE < total) {
-        await load_incrementally.call(this, view, schema, hidden, json, total, page + 1);
-    }
-}
-
 async function grid(div, view, hidden) {
     let [nrows, json, schema] = await Promise.all([
         view.num_rows(), 
@@ -943,14 +603,14 @@ async function grid(div, view, hidden) {
     }
     json.length = nrows;
     if (visible_rows.length > 0) {
-        json = await fill_page(view, json, hidden, visible_rows[1], visible_rows[visible_rows.length - 1] + 1);        
+        json = await fill_page(view, json, hidden, visible_rows[1], visible_rows[visible_rows.length - 1] + 2);        
     }
     if (!(document.contains ? document.contains(this.grid) : false)) {
         div.innerHTML = "";
         div.appendChild(this.grid);
     }
     this.grid.grid._cache_update = async (s, e) => {
-        json = await fill_page(view, json, hidden, s, e + 1);        
+        json = await fill_page(view, json, hidden, s, e + 1);  
         this.grid.set_data(json, schema);
     }
     if (visible_rows.length > 0) {
@@ -958,7 +618,6 @@ async function grid(div, view, hidden) {
         this.grid.grid.canvas.resize();
         this.grid.grid.canvas.resize();
     }
-   // await load_incrementally.call(this, view, schema, hidden, json, nrows, 0);
 }
 
 global.registerPlugin("hypergrid", {
